Moved tests to GameTests File
